Refactoring, Run GUI changes
Refactored functions that used the allRuns field to now use the DO2RunIterator class.
Added the '/runs' command [Admins Only] which server as an alias for '/decked-out gui'.
In the Run GUI, lackey runs now properly attribute the run wins to the finishers, not all of the runners.
Combined all the ItemStacks in bought_items in the database (so instead of having like 5 ItemStack of 2 victory tomes, you have one ItemStack of 10 tomes)