v.1.4.0
Runs that are not loaded from the database now properly show up in the Run GUI.
The current run info and queue are now saved more often, to prevent data loss if the server crashes.

Added the '/makephase' command, which makes the item you're holding into a phase item.

Added the '/cmd' command to alter/get the CustomModelData of the item you're holding.
'/cmd get' shows the CMD in chat.
'/cmd set <int>' sets the CMD of the item.